+++ src/live-share/test/suite/live-share.test.ts
@@ -3,7 +3,7 @@ import sinon = require('sinon');
 import assert = require('assert');
 import * as vsls from 'vsls';
 
-import { IPeacockSettings, Commands, ColorSettings, timeout } from '../../../models';
+import { IPeacockSettings, Commands, ColorSettings, timeout, azureBlue } from '../../../models';
 import {
   setupTestSuite,
   teardownTestSuite,
@@ -11,59 +11,56 @@ import {
 } from '../../../test/suite/lib/setup-teardown-test-suite';
 import { isValidColorInput } from '../../../color-library';
 import { executeCommand } from '../../../test/suite/lib/constants';
-import { LiveShareCommands } from '../../enums';
-import { peacockVslsMementos } from '../../constants';
-import { getPeacockWorkspaceConfig } from '../../../configuration';
+import { LiveShareCommands, LiveShareSettings } from '../../enums';
+import {
+  getPeacockWorkspaceConfig,
+  getLiveShareColor,
+  updateLiveShareColor,
+} from '../../../configuration';
 
 suite('Live Share Integration', () => {
   let originalValues = <IPeacockSettings>{};
+  let extensionContext: vscode.ExtensionContext | undefined;
 
   suiteSetup(async () => await setupTestSuite(originalValues));
   suiteTeardown(async () => await teardownTestSuite(originalValues));
   setup(async () => await setupTest());
 
-  test('can set color setting for Live Share host', async () => {
-    // Stub the async quick pick to return a response
-    const color = '#007fff';
-    const fakeResponse = `Azure Blue -> ${color}`;
-    const stub = await sinon
-      .stub(vscode.window, 'showQuickPick')
-      .returns(Promise.resolve<any>(fakeResponse));
-
-    const extensionContext = await executeCommand<vscode.ExtensionContext>(
+  setup(async () => {
+    extensionContext = await executeCommand<vscode.ExtensionContext>(
       LiveShareCommands.changeColorOfLiveShareHost,
     );
+  });
 
-    const settingValue =
-      (await extensionContext!.globalState.get<string>(peacockVslsMementos.vslsShareColor)) || '';
-    await extensionContext!.globalState.update(peacockVslsMementos.vslsShareColor, null);
+  teardown(async () => {
+    await updateLiveShareColor(LiveShareSettings.VSLSShareColor, '');
+  });
+
+  test('can set color setting for Live Share host', async () => {
+    // Stub the async quick pick to return a response
+    const fakeResponse = `Azure Blue -> ${azureBlue}`;
+    const stub = await stubQickPick(fakeResponse);
+
+    const settingValue = getLiveShareColor(LiveShareSettings.VSLSShareColor) || '';
 
     stub.restore();
 
     assert(isValidColorInput(settingValue));
-    assert(settingValue === color);
+    assert(settingValue === azureBlue);
   });
 
   test('can set color setting for Live Share guest', async () => {
     // Stub the async quick pick to return a response
-    const color = '#007fff';
-    const fakeResponse = `Azure Blue -> ${color}`;
-    const stub = await sinon
-      .stub(vscode.window, 'showQuickPick')
-      .returns(Promise.resolve<any>(fakeResponse));
+    const fakeResponse = `Azure Blue -> ${azureBlue}`;
+    const stub = await stubQickPick(fakeResponse);
 
-    const extensionContext = await executeCommand<vscode.ExtensionContext>(
-      LiveShareCommands.changeColorOfLiveShareGuest,
-    );
-
-    const settingValue =
-      (await extensionContext!.globalState.get<string>(peacockVslsMementos.vslsJoinColor)) || '';
-    await extensionContext!.globalState.update(peacockVslsMementos.vslsJoinColor, null);
+    const settingValue = getLiveShareColor(LiveShareSettings.VSLSJoinColor) || '';
+    await updateLiveShareColor(LiveShareSettings.VSLSJoinColor, '');
 
     stub.restore();
 
     assert(isValidColorInput(settingValue));
-    assert(settingValue === color);
+    assert(settingValue === azureBlue);
   });
 
   test('Workspace color is updated when Live Share session is started.', async () => {
@@ -73,15 +70,9 @@ suite('Live Share Integration', () => {
       throw new Error('Live Share extension is not installed.');
     }
 
-    const color = '#007fff';
-    const fakeResponse = `Azure Blue -> ${color}`;
-    const stub = await sinon
-      .stub(vscode.window, 'showQuickPick')
-      .returns(Promise.resolve<any>(fakeResponse));
+    const fakeResponse = `Azure Blue -> ${azureBlue}`;
+    const stub = await stubQickPick(fakeResponse);
 
-    const extensionContext = await executeCommand<vscode.ExtensionContext>(
-      LiveShareCommands.changeColorOfLiveShareHost,
-    );
     stub.restore();
 
     await vslsApi.share();
@@ -91,10 +82,9 @@ suite('Live Share Integration', () => {
     const value = config[ColorSettings.titleBar_activeBackground];
 
     await vslsApi.end();
-    await extensionContext!.globalState.update(peacockVslsMementos.vslsShareColor, null);
 
     assert(isValidColorInput(value));
-    assert(value === color);
+    assert(value === azureBlue);
   });
 
   test('Workspace color is reverted when Live Share session is ended.', async () => {
@@ -104,15 +94,9 @@ suite('Live Share Integration', () => {
       throw new Error('Live Share extension is not installed.');
     }
 
-    const color = '#007fff';
-    const fakeResponse = `Azure Blue -> ${color}`;
-    const stub = await sinon
-      .stub(vscode.window, 'showQuickPick')
-      .returns(Promise.resolve<any>(fakeResponse));
+    const fakeResponse = `Azure Blue -> ${azureBlue}`;
+    const stub = await stubQickPick(fakeResponse);
 
-    const extensionContext = await executeCommand<vscode.ExtensionContext>(
-      LiveShareCommands.changeColorOfLiveShareHost,
-    );
     stub.restore();
 
     await vslsApi.share();
@@ -123,7 +107,7 @@ suite('Live Share Integration', () => {
     let config = getPeacockWorkspaceConfig();
     const value = config[ColorSettings.titleBar_activeBackground];
 
-    await extensionContext!.globalState.update(peacockVslsMementos.vslsShareColor, null);
+    // await updateLiveShareColor(LiveShareSettings.VSLSShareColor, '');
 
     assert(!isValidColorInput(value));
     assert(value == null);
@@ -138,25 +122,18 @@ suite('Live Share Integration', () => {
 
     await vslsApi.share();
 
-    const color = '#007fff';
-    const fakeResponse = `Azure Blue -> ${color}`;
-    const stub = await sinon
-      .stub(vscode.window, 'showQuickPick')
-      .returns(Promise.resolve<any>(fakeResponse));
+    const fakeResponse = `Azure Blue -> ${azureBlue}`;
+    const stub = await stubQickPick(fakeResponse);
 
-    const extensionContext = await executeCommand<vscode.ExtensionContext>(
-      LiveShareCommands.changeColorOfLiveShareHost,
-    );
     stub.restore();
 
     let config = getPeacockWorkspaceConfig();
     const value = config[ColorSettings.titleBar_activeBackground];
 
     await vslsApi.end();
-    await extensionContext!.globalState.update(peacockVslsMementos.vslsShareColor, null);
 
     assert(isValidColorInput(value));
-    assert(value === color);
+    assert(value === azureBlue);
   });
 
   test('Workspace color is immediately reflected when updated during Live Share session.', async () => {
@@ -165,16 +142,9 @@ suite('Live Share Integration', () => {
     if (!vslsApi) {
       throw new Error('Live Share extension is not installed.');
     }
+    const fakeResponse = `Azure Blue -> ${azureBlue}`;
+    const stub = await stubQickPick(fakeResponse);
 
-    const color = '#007fff';
-    const fakeResponse = `Azure Blue -> ${color}`;
-    const stub = await sinon
-      .stub(vscode.window, 'showQuickPick')
-      .returns(Promise.resolve<any>(fakeResponse));
-
-    const extensionContext = await executeCommand<vscode.ExtensionContext>(
-      LiveShareCommands.changeColorOfLiveShareHost,
-    );
     stub.restore();
 
     await vslsApi.share();
@@ -183,9 +153,7 @@ suite('Live Share Integration', () => {
 
     const color2 = '#68217a';
     const fakeResponse2 = `C# Purple -> ${color2}`;
-    const stub2 = await sinon
-      .stub(vscode.window, 'showQuickPick')
-      .returns(Promise.resolve<any>(fakeResponse2));
+    const stub2 = await stubQickPick(fakeResponse2);
 
     await executeCommand<vscode.ExtensionContext>(LiveShareCommands.changeColorOfLiveShareHost);
     stub2.restore();
@@ -199,7 +167,7 @@ suite('Live Share Integration', () => {
     assert(value === color2);
 
     await vslsApi.end();
-    await extensionContext!.globalState.update(peacockVslsMementos.vslsShareColor, null);
+    // await updateLiveShareColor(LiveShareSettings.VSLSShareColor, '');
   });
 
   test('Workspace color is reverted after Live Share session when updated the color multiple times.', async () => {
@@ -215,22 +183,15 @@ suite('Live Share Integration', () => {
 
     const color2 = '#68217a';
     const fakeResponse2 = `C# Purple -> ${color2}`;
-    const stub2 = await sinon
-      .stub(vscode.window, 'showQuickPick')
-      .returns(Promise.resolve<any>(fakeResponse2));
+    const stub2 = await stubQickPick(fakeResponse2);
 
-    const extensionContext = await executeCommand<vscode.ExtensionContext>(
-      LiveShareCommands.changeColorOfLiveShareHost,
-    );
     stub2.restore();
 
     await timeout(1000);
 
     const color3 = '#b52e31';
-    const fakeResponse3 = `Abgular Red -> ${color3}`;
-    const stub3 = await sinon
-      .stub(vscode.window, 'showQuickPick')
-      .returns(Promise.resolve<any>(fakeResponse3));
+    const fakeResponse3 = `Angular Red -> ${color3}`;
+    const stub3 = await stubQickPick(fakeResponse3);
 
     await executeCommand<vscode.ExtensionContext>(LiveShareCommands.changeColorOfLiveShareHost);
     stub3.restore();
@@ -239,9 +200,7 @@ suite('Live Share Integration', () => {
 
     const color4 = '#639';
     const fakeResponse4 = `Gatsby Purple -> ${color4}`;
-    const stub4 = await sinon
-      .stub(vscode.window, 'showQuickPick')
-      .returns(Promise.resolve<any>(fakeResponse4));
+    const stub4 = await stubQickPick(fakeResponse4);
 
     await executeCommand<vscode.ExtensionContext>(LiveShareCommands.changeColorOfLiveShareHost);
     stub4.restore();
@@ -258,7 +217,7 @@ suite('Live Share Integration', () => {
     assert(!isValidColorInput(value));
     assert(value == null);
 
-    await extensionContext!.globalState.update(peacockVslsMementos.vslsShareColor, null);
+    // await updateLiveShareColor(LiveShareSettings.VSLSShareColor, '');
   });
 
   test('Workspace color is reverted to a preset color after Live Share session when updated the color multiple times.', async () => {
@@ -268,11 +227,9 @@ suite('Live Share Integration', () => {
       throw new Error('Live Share extension is not installed.');
     }
 
-    const startColor = '#007fff';
+    const startColor = azureBlue;
     const fakeResponse = `Azure Blue -> ${startColor}`;
-    const stub = await sinon
-      .stub(vscode.window, 'showQuickPick')
-      .returns(Promise.resolve<any>(fakeResponse));
+    const stub = await stubQickPick(fakeResponse);
 
     await executeCommand(Commands.changeColorToFavorite);
     stub.restore();
@@ -283,22 +240,15 @@ suite('Live Share Integration', () => {
 
     const color2 = '#68217a';
     const fakeResponse2 = `C# Purple -> ${color2}`;
-    const stub2 = await sinon
-      .stub(vscode.window, 'showQuickPick')
-      .returns(Promise.resolve<any>(fakeResponse2));
+    const stub2 = await stubQickPick(fakeResponse2);
 
-    const extensionContext = await executeCommand<vscode.ExtensionContext>(
-      LiveShareCommands.changeColorOfLiveShareHost,
-    );
     stub2.restore();
 
     await timeout(1000);
 
     const color3 = '#b52e31';
     const fakeResponse3 = `Abgular Red -> ${color3}`;
-    const stub3 = await sinon
-      .stub(vscode.window, 'showQuickPick')
-      .returns(Promise.resolve<any>(fakeResponse3));
+    const stub3 = await stubQickPick(fakeResponse3);
 
     await executeCommand<vscode.ExtensionContext>(LiveShareCommands.changeColorOfLiveShareHost);
     stub3.restore();
@@ -307,9 +257,7 @@ suite('Live Share Integration', () => {
 
     const color4 = '#639';
     const fakeResponse4 = `Gatsby Purple -> ${color4}`;
-    const stub4 = await sinon
-      .stub(vscode.window, 'showQuickPick')
-      .returns(Promise.resolve<any>(fakeResponse4));
+    const stub4 = await stubQickPick(fakeResponse4);
 
     await executeCommand<vscode.ExtensionContext>(LiveShareCommands.changeColorOfLiveShareHost);
     stub4.restore();
@@ -326,6 +274,13 @@ suite('Live Share Integration', () => {
     assert(isValidColorInput(value));
     assert(value === startColor);
 
-    await extensionContext!.globalState.update(peacockVslsMementos.vslsShareColor, null);
+    // await updateLiveShareColor(LiveShareSettings.VSLSShareColor, '');
   });
 });
+
+const stubQickPick = async (fakeResponse: string) => {
+  const stub = await sinon
+    .stub(vscode.window, 'showQuickPick')
+    .returns(Promise.resolve<any>(fakeResponse));
+  return stub;
+};