+++ src/models/enums.ts
@@ -1,4 +1,5 @@
 import { RemoteSettings } from '../remote/enums';
+import { LiveShareSettings } from '../live-share/enums';
 
 export enum StandardSettings {
   ElementAdjustments = 'elementAdjustments',
@@ -21,7 +22,7 @@ export enum AffectedSettings {
   TabActiveBorder = 'affectTabActiveBorder',
 }
 
-export type AllSettings = StandardSettings | AffectedSettings | RemoteSettings;
+export type AllSettings = StandardSettings | AffectedSettings | RemoteSettings | LiveShareSettings;
 
 export enum Commands {
   resetColors = 'peacock.resetColors',